Add WED status updates on LCD
Indicate if door lock has WED status 'disabled', 'enabled', 'linking', or 'linked':
Disabled : Door lock is not sampling for Wake-up frames.
Enabled  : Door lock is sampling for Wake-up frames.
Linking  : Door lock is in the process of attaching / linking to the Wake-up Coordinator.
Linked   : Attached to the Wake-up Coordinator and ready to communicate over an eCSL link.